refactor: Pin 구조 변경
