--- src/main/java/bowling/domain/pin/Pin.java
@@ -1,6 +1,0 @@
-package bowling.domain.pin;
-
-public interface Pin {
-
-    int getPin();
-}
