+++ src/test/java/bowling/domain/PinTest.java
@@ -1,4 +1,4 @@
-package bowling.domain.pin;
+package bowling.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -6,14 +6,14 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class NormalPinTest {
+class PinTest {
 
     @DisplayName("범위를 초과하는 핀을 쓰러뜨리면 예외가 발생한다.")
     @ParameterizedTest
     @ValueSource(ints = {-1, 11})
     void valid_pin(int countOfDownPin) {
         assertThatThrownBy(() -> {
-            NormalPin.from(countOfDownPin);
+            Pin.from(countOfDownPin);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 }
